Eclipse project file
This file contains the code necessary for training a neural network given a series of inputs, and interpreting new inputs after it has been trained. The neural network contains 30*30 = 900 inputs that are connected to 2 neuron groups with 8 neurons each, which are then connected to two outputs. The code in the project is designed to learn how to differenciate a triangle from a rectangle using the absolue brute force method (lol), and can achieve a 67.5% accuracy. Higher accuracies can also be achieved with a much longer training time.